backend/requirements.txt
django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.6
gunicorn==20.1.0 
celery==5.3.6
redis==4.6.0
drf-yasg==1.21.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-beat==2.5.0
flower==2.0.1
pytest==8.3.5
pytest-django==4.11.1

